requirements.txt
google-auth-oauthlib>=0.4.6
google-auth>=2.3.3
google-api-python-client>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.0
aiofiles>=0.8.0
cryptography>=3.4.7
python-dotenv>=0.19.0
pytest>=6.0.0
pytest-asyncio>=0.15.0
pytest-cov>=2.12.0